# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface MobileTokenExchangeRequest {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  code_verifier: string;
  /** @minLength 1 */
  redirect_uri: string;
  /** @minLength 1 */
  state: string;
  /** @minLength 1 */
  nonce?: string;
}

export interface MobileTokenExchangeSuccess {
  token: string;
}

export const LogoutSuccessValue = {
  success: true,
} as const;
export type LogoutSuccess = typeof LogoutSuccessValue;

export interface ErrorEnvelope {
  error: string;
}

export interface Clinic {
  id: string;
  slug: string;
  name: string;
  doctorName: string;
  avgConsultationMinutes: number;
  /** E.164 style phone number, no plus sign required */
  whatsappNumber: string;
  createdAt: string;
}

export interface ClinicEnvelope {
  clinic: Clinic | null;
}

export interface CreateClinicRequest {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  doctorName: string;
  /**
   * @minimum 1
   * @maximum 120
   */
  avgConsultationMinutes: number;
  /** @minLength 6 */
  whatsappNumber: string;
}

export interface UpdateClinicRequest {
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  doctorName?: string;
  /**
   * @minimum 1
   * @maximum 120
   */
  avgConsultationMinutes?: number;
  /** @minLength 6 */
  whatsappNumber?: string;
}

export interface ClinicStats {
  totalToday: number;
  served: number;
  waiting: number;
  inProgress: number;
  skipped: number;
  avgWaitMinutes: number;
  /** @nullable */
  currentTokenNumber: number | null;
  /** @nullable */
  nextTokenNumber: number | null;
}

export type PatientStatus = (typeof PatientStatus)[keyof typeof PatientStatus];

export const PatientStatus = {
  waiting: "waiting",
  in_progress: "in_progress",
  done: "done",
  skipped: "skipped",
} as const;

export type PatientReminderStage =
  (typeof PatientReminderStage)[keyof typeof PatientReminderStage];

export const PatientReminderStage = {
  none: "none",
  three_away: "three_away",
  two_away: "two_away",
  one_away: "one_away",
  your_turn: "your_turn",
  done: "done",
} as const;

export interface Patient {
  id: string;
  clinicId: string;
  name: string;
  phone: string;
  tokenNumber: number;
  status: PatientStatus;
  trackingCode: string;
  /** 0 = currently being served, 1 = next, etc. */
  position: number;
  estimatedWaitMinutes: number;
  reminderStage: PatientReminderStage;
  createdAt: string;
  /** @nullable */
  completedAt?: string | null;
}

export interface AddPatientRequest {
  /** @minLength 1 */
  name: string;
  /** @minLength 4 */
  phone: string;
}

export interface AdvanceQueueResult {
  completed: Patient | null;
  current: Patient | null;
}

export interface PublicClinic {
  id: string;
  slug: string;
  name: string;
  doctorName: string;
  whatsappNumber: string;
  avgConsultationMinutes: number;
}

export interface PublicJoinRequest {
  /** @minLength 1 */
  name: string;
  /** @minLength 4 */
  phone: string;
}

export interface PublicJoinResponse {
  clinicName: string;
  doctorName: string;
  tokenNumber: number;
  estimatedWaitMinutes: number;
  position: number;
  trackingCode: string;
  trackingUrl: string;
  whatsappConfirmationMessage: string;
}

export type PublicTrackingStatus =
  (typeof PublicTrackingStatus)[keyof typeof PublicTrackingStatus];

export const PublicTrackingStatus = {
  waiting: "waiting",
  in_progress: "in_progress",
  done: "done",
  skipped: "skipped",
} as const;

export type PublicTrackingReminderStage =
  (typeof PublicTrackingReminderStage)[keyof typeof PublicTrackingReminderStage];

export const PublicTrackingReminderStage = {
  none: "none",
  three_away: "three_away",
  two_away: "two_away",
  one_away: "one_away",
  your_turn: "your_turn",
  done: "done",
} as const;

export interface PublicTracking {
  clinicName: string;
  doctorName: string;
  tokenNumber: number;
  status: PublicTrackingStatus;
  position: number;
  patientsAhead: number;
  estimatedWaitMinutes: number;
  /** @nullable */
  currentTokenNumber: number | null;
  reminderStage: PublicTrackingReminderStage;
  avgConsultationMinutes: number;
}

/**
 * Opaque session token — `Bearer <sid>`.
 */
export type AuthorizationSessionHeaderParameter = string;

export type BeginBrowserLoginParams = {
  /**
   * Relative path to redirect to after login (must start with `/`). Defaults to `/`.
   */
  returnTo?: string;
};

export type HandleBrowserLoginCallbackParams = {
  code?: string;
  state?: string;
  iss?: string;
};